More work to revive elimination tournaments and quick_game.

+++ app/src/main/java/com/twobits/pocketleague/db/tables/Game.java
@@ -55,7 +55,7 @@ public class Game extends CouchDocumentBase {
     }
 
     // Other methods
-    public long getIdInSession() {
+    public int getIdInSession() {
         return (int) content.get(ID_IN_SESSION);
     }
